+++ client/logmon/logging/rotator.go
