+++ client/logmon/logging/rotator_test.go
