+++ client/logmon/logging/syslog_parser.go
