+++ client/logmon/logging/syslog_server.go
