+++ client/logmon/logmon.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	hclog "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/nomad/client/driver/logging"
 	"github.com/hashicorp/nomad/client/lib/fifo"
+	"github.com/hashicorp/nomad/client/logmon/logging"
 )
 
 const (
